+++ .changeset/slow-radios-rush.md
@@ -1,0 +1,5 @@
+---
+"@evmts/core": patch
+---
+
+Fixed bug with `args` key being undefined rather than not existing in Object.keys
