+++ core/src/index.ts
@@ -1,1 +1,2 @@
-export * from './contract'
+export * from './EvmtsContract'
+export * from './evmtsContractFactory'
